refactor: redis 관련 메소드들 분리
