+++ src/main/java/zoo/insightnote/domain/payment/service/KakaoPayService.java
@@ -26,94 +26,16 @@ import java.util.concurrent.TimeUnit;
 @RequiredArgsConstructor
 @Slf4j
 public class KakaoPayService {
-
     private final RestTemplate restTemplate;
     private final ObjectMapper objectMapper;
-    private final RedisTemplate<String, String> redisTemplate;
-    private final long PAYMENT_TID_EXPIRATION = 10 * 60; // 10분
-    private final long PAYMENT_SESSION_KEYS_EXPIRATION = 5 * 60; // 5분
+    private final PaymentRedisService paymentRedisService;
+
 
     @Value("${kakao.api.cid}")
     private String cid;
 
     @Value("${kakao.api.admin-key}")
     private String adminKey;
-
-
-    private void saveTidKey(Long orderId, String tid) {
-        String tidKey = "payment:tid: " + orderId;
-        redisTemplate.opsForValue().set(tidKey, tid, PAYMENT_TID_EXPIRATION, TimeUnit.SECONDS);
-    }
-
-    public String getTidKey(Long orderId) {
-        String tidKey = "payment:tid: " + orderId;
-        String tid = redisTemplate.opsForValue().get(tidKey);
-        if (tid == null) {
-            log.error("❌ Redis에서 tid 조회 실패! (orderId={})", orderId);
-            throw new CustomException(ErrorCode.PAYMENT_NOT_FOUND);
-        }
-        return tid;
-    }
-
-    private void saveSessionIds(Long orderId, List<Long> sessionsId) {
-        String sessionIdsKey = "payment:sessions: " + orderId;
-        try {
-            // ✅ JSON으로 변환하여 Redis에 저장
-            String jsonSessionIds = objectMapper.writeValueAsString(sessionsId);
-            redisTemplate.opsForValue().set(sessionIdsKey, jsonSessionIds, PAYMENT_SESSION_KEYS_EXPIRATION, TimeUnit.SECONDS);
-        } catch (JsonProcessingException e) {
-            log.error("❌ JSON 변환 오류 (sessionIds 저장 실패)", e);
-            throw new CustomException(ErrorCode.JSON_PROCESSING_ERROR);
-        }
-    }
-
-    public List<Long> getSessionIds(Long orderId) {
-        String sessionIdsKey = "payment:sessions: " + orderId;
-        String getSessionsIds = redisTemplate.opsForValue().get(sessionIdsKey);
-
-        if (getSessionsIds == null) {
-            log.error("❌ Redis에서 sessions 조회 실패! (orderId={})", orderId);
-            throw new CustomException(ErrorCode.PAYMENT_NOT_FOUND);
-        }
-
-        try {
-            ObjectMapper objectMapper = new ObjectMapper();
-            return objectMapper.readValue(getSessionsIds, new TypeReference<List<Long>>() {});
-        } catch (JsonProcessingException e) {
-            log.error("❌ JSON 변환 오류 (sessionIds 조회 실패)", e);
-            throw new CustomException(ErrorCode.JSON_PROCESSING_ERROR);
-        }
-    }
-
-    public void saveUserInfo(Long orderId, UserInfoDto userInfo) {
-        String userInfoKey = "payment:userInfo: " + orderId;
-        try {
-            String jsonUserInfo = objectMapper.writeValueAsString(userInfo);
-            redisTemplate.opsForValue().set(userInfoKey, jsonUserInfo, PAYMENT_SESSION_KEYS_EXPIRATION, TimeUnit.SECONDS);
-        } catch (JsonProcessingException e) {
-            log.error("❌ JSON 변환 오류 (sessionIds 저장 실패)", e);
-            throw new CustomException(ErrorCode.JSON_PROCESSING_ERROR);
-        }
-    }
-
-    public UserInfoDto getUserInfo(Long orderId) {
-        String userInfoKey = "payment:userInfo: " + orderId;
-        String getUserInfo = redisTemplate.opsForValue().get(userInfoKey);
-
-        if (getUserInfo == null) {
-            log.error("❌ Redis에서 sessions 조회 실패! (orderId={})", orderId);
-            throw new CustomException(ErrorCode.PAYMENT_NOT_FOUND);
-        }
-
-        try {
-            ObjectMapper objectMapper = new ObjectMapper();
-            return objectMapper.readValue(getUserInfo, UserInfoDto.class);
-        } catch (JsonProcessingException e) {
-            log.error("❌ JSON 변환 오류 (userInfo 조회 실패)", e);
-            throw new CustomException(ErrorCode.JSON_PROCESSING_ERROR);
-        }
-    }
-
 
     // 결제 요청
     public ResponseEntity<KakaoPayReadyResponseDto> requestKakaoPayment(PaymentRequestReadyDto requestDto) {
@@ -132,9 +54,9 @@ public class KakaoPayService {
             String tid = response.getBody().getTid();
             log.info("✅ 카카오페이 결제 요청 성공");
 
-            saveTidKey(orderId, tid);
-            saveSessionIds(orderId, requestDto.getSessionIds());
-            saveUserInfo(orderId, requestDto.getUserInfo());
+            paymentRedisService.saveTidKey(orderId, tid);
+            paymentRedisService.saveSessionIds(orderId, requestDto.getSessionIds());
+            paymentRedisService.saveUserInfo(orderId, requestDto.getUserInfo());
 
             return response;
         } catch (Exception e) {
